Melhorando as validações

+++ src/main/java/com/lmmartins/vrum/services/MotoristaService.java
@@ -1,19 +1,14 @@
 package com.lmmartins.vrum.services;
 
-import com.lmmartins.vrum.dto.CorridaDTO;
 import com.lmmartins.vrum.dto.MotoristaDTO;
 import com.lmmartins.vrum.enums.MotoristaStatus;
 import com.lmmartins.vrum.exceptions.ValidacaoException;
 import com.lmmartins.vrum.models.Motorista;
-import com.lmmartins.vrum.models.Passageiro;
 import com.lmmartins.vrum.repositories.MotoristaRepository;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Example;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.PathVariable;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -36,39 +31,49 @@ public class MotoristaService {
         return repository.findById(motoristaId);
     }
 
-    public Optional<Motorista> getMotoristaByCpf(String cpf) {
-        return repository.getMotoristaByCpf(cpf);
-    }
     //Metódos de Criação
     public Long criarMotorista(MotoristaDTO motoristaDto) throws Exception {
-        validarMotorista( motoristaDto);
         Motorista motorista = mapper.map(motoristaDto, Motorista.class);
+        validarMotorista(motorista);
         return repository.save(motorista).getId();
     }
 
     //Metódos de Atualização
     public Long atualizarMotorista(Long motoristaId,
                                    MotoristaDTO motoristaDto) throws Exception {
-        validarMotorista(motoristaDto);
-        Motorista motorista = mapper.map(motoristaDto, Motorista.class);
-        motorista.setId(motoristaId);
-        return repository.save(motorista).getId();
+        if(repository.existsById(motoristaId)) {
+            Motorista motorista = mapper.map(motoristaDto, Motorista.class);
+            validarMotorista(motorista);
+            motorista.setId(motoristaId);
+            return repository.save(motorista).getId();
+        } else {
+            throw new ValidacaoException("Motorista não existe.");
+        }
     }
 
-    public Integer atualizarStatusDoMotorista(Long motoristaId, MotoristaStatus status) {
-        return repository.setMotoristaStatus(motoristaId, status.motoristaStatus);
+    public void atualizarStatusDoMotorista(Long motoristaId, MotoristaStatus status) throws Exception {
+         if (!repository.existsById(motoristaId)) {
+             throw new ValidacaoException("Motorista não existe");
+         }
+        repository.setMotoristaStatus(motoristaId, status.motoristaStatus);
     }
 
     //Metódos de Deleção.
-    public void deletarMotoristaPorId(@PathVariable("motoristaId") Long motoristaId) {
+    public void deletarMotoristaPorId(Long motoristaId) throws Exception {
+        if (!repository.existsById(motoristaId)) {
+            throw new ValidacaoException("Motorista não existe");
+        }
         repository.deleteById(motoristaId);
     }
 
-    private void validarMotorista(MotoristaDTO motoristaDto) throws Exception {
-        Optional<Motorista> motorista= repository.getMotoristaByCpf(motoristaDto.getCpf());
+    //Validações
+    private void validarMotorista(Motorista motorista) throws Exception {
+        //TODO: Validar outros campos.
+    }
 
-        if(motorista.isPresent()) {
-            throw new ValidacaoException("Motorista já existe.");
+    private void validarMotoristaPorCPF(String cpf) throws Exception {
+        if(repository.existsByCpf(cpf)) {
+            throw new ValidacaoException("CPF já existe.");
         }
     }
 }
